requirements.txt
attrs==21.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cupy-cuda114==9.6.0
cymem==2.0.6
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.2.0/en_core_web_trf-3.2.0-py3-none-any.whl
fastrlock==0.8
filelock==3.4.0
huggingface-hub==0.2.1
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
jsonlines==3.0.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.6
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
srsly==2.4.2
thinc==8.0.13
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
transformers==4.12.5
typer==0.4.0
urllib3==1.26.7
utils-package-Martin36==0.0.6
wasabi==0.9.0
